Obligatorio1_Arancet_Cohen/Obligatorio1_Arancet_Cohen/Client.cs
using System;

namespace Obligatorio1ArancetCohen
{
    public class Client
    {
        public string Name { get; set; }
        public string Surname { get; set; }
        public string UserName { get; set; }
        public string Password { get; set; }
        public string Phone { get; set; }
        public string Id { get; set; }
        public DateTime RegistrationDate { get; set; }
        public DateTime LastLoginDate { get; set; }

        public Client(string name, string surname, string userName, string password, string phone, string id, DateTime registrationDate)
        {
            Name = name;
            Surname = surname;
            UserName = userName;
            Password = password;
            Phone = phone;
            Id = id;
            RegistrationDate = registrationDate;
            LastLoginDate = Constants.NEVER;
        }

        public DateTime updateLastLoginDate()
        {
            DateTime dateOfLogin = DateTime.Now;
            LastLoginDate = dateOfLogin;
            return dateOfLogin;
        }
    }

    
}